requirements.txt
-i https://pypi.org/simple
asgiref==3.6.0 ; python_version >= '3.7'
django==4.1.3
django-environ==0.9.0
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
markdown==3.4.1
pillow==9.3.0
pyjwt==2.6.0 ; python_version >= '3.7'
pytz==2022.7
sqlparse==0.4.3 ; python_version >= '3.5'
openpyxl=='3.0.10'
